getInnerValues
